requirements.txt
python-dotenv
azure-search-documents
azure-identity
msgraph-sdk